backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==2.3.7
